refactor: Refactor update_competition_json to automatically update the clubs list as well to have a unique function call
This changes the update_competition_json function which used to takes only update the json file to update both the list currently loaded and the competitions json, allowing for a unique function call.

The separation didn't make sense as updating the list was necessary to update the json in the first place.