refactor: Refactor club_json_update to also updating the competitions list to allow for a unique function calls
This changes the update_club_json function in repository/club which used to only update the json file to update both the list of clubs currently loaded and the clubs json, allowing for a unique function call.

The separation didn't make sense as updating the list was necessary to update the json in the first place.