fix: Add default value to the places value to avoid error
This add a default to the places value which used to send an empty if no number was given, raising an error.

BREAKING CHANGES
booking.html used to have not default value as places, which caused the initial value to be an empty string, raising an error since value expected is an integer (as string). Now have an initial value of 0.