refactor: Add models instances to the arguments for creation to avoid creating new ones in methods
This changes the BookingHandler class, adding a ClubService and a Competition services instances as arguments to be passed.

Previously created instance in methods.

BREAKING CHANGES
The class now declares the passed ClubService and CompetitionService instance as attributes.

The find_booking_data method used to create instances of ClubService and CompetitionService, now uses class attributes instead.