refactor: Declare filename outside function to allow easier change/testing
In repository, both file had a json reader and a json writer, which used to have the name of targeted json directly writing when calling the method open.

Now added a new variable outside the function, set up to be the filename. The method now use the variable as arg.

As such:
- Old: with open('filename.json')
- New: with open(var_filename)